Add registerCommand() and mgRegisterShapeCreator().
Change the default tolerance.
Use OSAtomicIncrement32() on iOS.
Add drawImage() in the GiCanvasIos class.

+++ core/include/graph/gicontxt.h
@@ -41,13 +41,13 @@ public:
         \param style 线型，取值为 kLineSolid 等
         \param fillcr 填充颜色， GiColor::Invalid() 表示不填充
     */
-    GiContext(Int16 width, GiColor color = GiColor::Black(), 
+    GiContext(float width, GiColor color = GiColor::Black(), 
               kLineStyle style = kLineSolid, const GiColor& fillcr = GiColor::Invalid())
         : m_type(0), m_lineStyle(style), m_lineWidth(width)
         , m_lineColor(color), m_fillColor(fillcr)
     {
     }
-
+    
     //! 拷贝构造函数
     GiContext(const GiContext& src)
     {
@@ -56,7 +56,7 @@ public:
         m_lineColor = src.m_lineColor;
         m_fillColor = src.m_fillColor;
     }
-
+    
     //! 赋值操作符函数
     GiContext& operator=(const GiContext& src)
     {
@@ -69,7 +69,7 @@ public:
         }
         return *this;
     }
-
+    
     //! 比较相等操作符函数
     bool operator==(const GiContext& src) const
     {
@@ -78,25 +78,25 @@ public:
             && m_lineColor == src.m_lineColor
             && m_fillColor == src.m_fillColor;
     }
-
+    
     //! 比较不相等操作符函数
     bool operator!=(const GiContext& src) const
     {
         return !operator==(src);
     }
-
+    
     //! 返回线型
     kLineStyle getLineStyle() const
     {
         return m_lineColor.isInvalid() ? kLineNull : m_lineStyle;
     }
-
+    
     //! 设置线型
     void setLineStyle(kLineStyle style)
     {
         m_lineStyle = style;
     }
-
+    
     //! 返回线宽
     /*! 正数表示单位为0.01mm，零表示1像素宽，负数时表示单位为像素
     */
@@ -104,7 +104,7 @@ public:
     {
         return m_lineWidth;
     }
-
+    
     //! 设置线宽
     /*!
         \param width 线宽，正数表示单位为0.01mm，零表示1像素宽，负数时表示单位为像素
@@ -113,13 +113,13 @@ public:
     {
         m_lineWidth = width;
     }
-
+    
     //! 返回是否为空线，即不画线
     bool isNullLine() const
     {
         return m_lineStyle == kLineNull || m_lineColor.isInvalid();
     }
-
+    
     //! 设置为空线，即不画线
     /*! 如果要恢复成普通画线状态，可调setLineStyle(kLineSolid)
     */
@@ -127,67 +127,67 @@ public:
     {
         m_lineStyle = kLineNull;
     }
-
+    
     //! 返回线条颜色
     GiColor getLineColor() const
     {
         return m_lineColor;
     }
-
+    
     //! 设置线条颜色， GiColor::Invalid() 表示不画线条
     void setLineColor(const GiColor& color)
     {
         m_lineColor = color;
     }
-
+    
     //! 返回线条透明度
     UInt8 getLineAlpha() const
     {
         return m_lineColor.a;
     }
-
+    
     //! 设置线条透明度，0到255，0表示全透明，255表示不透明
     void setLineAlpha(UInt8 alpha)
     {
         m_lineColor.a = alpha;
     }
-
+    
     //! 返回是否填充
     bool hasFillColor() const
     {
         return !m_fillColor.isInvalid();
     }
-
+    
     //! 设置为不填充
     void setNoFillColor()
     {
         m_fillColor = GiColor::Invalid();
     }
-
+    
     //! 返回填充颜色
     GiColor getFillColor() const
     {
         return m_fillColor;
     }
-
+    
     //! 设置填充颜色， GiColor::Invalid() 表示不填充
     void setFillColor(const GiColor& color)
     {
         m_fillColor = color;
     }
-
+    
     //! 返回填充透明度
     UInt8 getFillAlpha() const
     {
         return m_fillColor.a;
     }
-
+    
     //! 设置填充透明度，0到255，0表示全透明，255表示不透明
     void setFillAlpha(UInt8 alpha)
     {
         m_fillColor.a = alpha;
     }
-
+    
     //! 返回绘图环境类型，供派生类用
     /*! 本类始终返回1，派生类可设置 m_type 为其他值
     */
@@ -195,7 +195,7 @@ public:
     {
         return m_type;
     }
-
+    
 protected:
     int         m_type;            //!< 派生类可指定其他值来表示不同类型
 private:
