+++ core/include/shape/mgshapet.h
@@ -12,7 +12,7 @@
 
 //! 矢量图形模板类
 /*! \ingroup GEOM_SHAPE
-*/
+ */
 template <class ShapeT, class ContextT = GiContext>
 class MgShapeT : public MgShape
 {
@@ -22,15 +22,15 @@ public:
     ContextT    _context;
     UInt32      _id;
     MgShapes*   _parent;
-
+    
     MgShapeT() : _id(0), _parent(NULL)
     {
     }
-
+    
     virtual ~MgShapeT()
     {
     }
-
+    
     GiContext* context()
     {
         return &_context;
@@ -40,52 +40,52 @@ public:
     {
         return &_context;
     }
-
+    
     MgBaseShape* shape()
     {
         return &_shape;
     }
-
+    
     const MgBaseShape* shape() const
     {
         return &_shape;
     }
-
+    
     bool draw(GiGraphics& gs, const GiContext *ctx = NULL) const
     {
         ContextT tmpctx(getContext(gs, ctx));
         return shape()->draw(gs, tmpctx);
     }
-
+    
     static MgShape* create()
     {
         return new ThisClass;
     }
-
+    
     static UInt32 Type() { return 10000 + ShapeT::Type(); }
     UInt32 getType() const { return Type(); }
-
+    
     bool isKindOf(UInt32 type) const
     {
         return type == Type() || type == MgShape::Type();
     }
-
+    
     void release()
     {
         delete this;
     }
-
+    
     MgObject* clone() const
     {
         ThisClass *p = new ThisClass;
-
+        
         p->shape()->copy(_shape);
         p->shape()->update();
         p->_context = _context;
-
+        
         return p;
     }
-
+    
     void copy(const MgObject& src)
     {
         if (src.isKindOf(Type())) {
@@ -101,26 +101,26 @@ public:
     bool equals(const MgObject& src) const
     {
         bool ret = false;
-
+        
         if (src.isKindOf(Type())) {
             const ThisClass& _src = (const ThisClass&)src;
             ret = shape()->equals(_src._shape)
-                && _context == _src._context;
+            && _context == _src._context;
         }
-
+        
         return ret;
     }
-
+    
     UInt32 getID() const
     {
         return _id;
     }
-
+    
     MgShapes* getParent() const
     {
         return _parent;
     }
-
+    
     void setParent(MgShapes* p, UInt32 nID)
     {
         _parent = p;
@@ -150,10 +150,15 @@ public:
         _context.setLineWidth(s->readFloat("lineWidth"));
         
         c = s->readUInt32("lineColor");
-        _context.setLineColor(GiColor(c & 0xFF, (c >> 8) & 0xFF, (c >> 16) & 0xFF, (c >> 24) & 0xFF));
-        
+        _context.setLineColor(GiColor((UInt8)(c & 0xFF), 
+                                      (UInt8)((c >> 8 ) & 0xFF), 
+                                      (UInt8)((c >> 16) & 0xFF), 
+                                      (UInt8)((c >> 24) & 0xFF)));
         c = s->readUInt32("fillColor");
-        _context.setFillColor(GiColor(c & 0xFF, (c >> 8) & 0xFF, (c >> 16) & 0xFF, (c >> 24) & 0xFF));
+        _context.setFillColor(GiColor((UInt8)(c & 0xFF), 
+                                      (UInt8)((c >> 8 ) & 0xFF), 
+                                      (UInt8)((c >> 16) & 0xFF), 
+                                      (UInt8)((c >> 24) & 0xFF)));
         
         bool ret = shape()->load(s);
         if (ret) {
@@ -167,7 +172,7 @@ protected:
     ContextT getContext(GiGraphics& gs, const GiContext *ctx) const
     {
         ContextT tmpctx(_context);
-
+        
         if (ctx && !ctx->isNullLine()) {
             float addw  = ctx->getLineWidth();
             float width = tmpctx.getLineWidth();
@@ -178,16 +183,16 @@ protected:
             else
                 tmpctx.setLineWidth(-addw);         // 换成新的像素宽度
         }
-
+        
         if (ctx && !ctx->isNullLine())
             tmpctx.setLineColor(ctx->getLineColor());
-
+        
         if (ctx && !ctx->isNullLine())
             tmpctx.setLineStyle(ctx->getLineStyle());
-
+        
         if (ctx && ctx->hasFillColor())
             tmpctx.setFillColor(ctx->getFillColor());
-
+        
         return tmpctx;
     }
 };
